+++ lib/resque/log_formatters/quiet_formatter.rb
@@ -1,0 +1,7 @@
+module Resque
+  class QuietFormatter
+    def call(serverity, datetime, progname, msg)
+      ""
+    end
+  end
+end