+++ test/worker_test.rb
@@ -476,22 +476,82 @@ describe "Resque::Worker" do
     assert !$BEFORE_FORK_CALLED, "before_fork should not have been called after job runs"
   end
 
+  it "setting verbose to true" do
+    @worker.verbose = true
+
+    assert @worker.verbose
+    assert !@worker.very_verbose
+  end
+
+  it "setting verbose to false" do
+    @worker.verbose = false
+
+    assert !@worker.verbose
+    assert !@worker.very_verbose
+  end
+
+  it "setting very_verbose to true" do
+    @worker.very_verbose = true
+
+    assert !@worker.verbose
+    assert @worker.very_verbose
+  end
+
+  it "setting setting verbose to true and then very_verbose to false" do
+    $warned_logger_severity_deprecation = true
+    @worker.very_verbose = true
+    @worker.verbose      = true
+    @worker.very_verbose = false
+
+    assert @worker.verbose
+    assert !@worker.very_verbose
+  end
+
+  it "verbose prints out logs" do
+    messages        = StringIO.new
+    Resque.logger   = Logger.new(messages)
+    @worker.verbose = true
+
+    begin
+      @worker.log("omghi mom")
+    ensure
+      reset_logger
+    end
+
+    assert_equal "*** omghi mom", messages.string
+  end
+
+  it "unsetting verbose works" do
+    messages        = StringIO.new
+    Resque.logger   = Logger.new(messages)
+    @worker.verbose = true
+    @worker.verbose = false
+
+    begin
+      @worker.log("omghi mom")
+    ensure
+      reset_logger
+    end
+
+    assert_equal "", messages.string
+  end
+
   it "very verbose works in the afternoon" do
+    messages        = StringIO.new
+    Resque.logger   = Logger.new(messages)
+
     begin
       require 'time'
       last_puts = ""
       Time.fake_time = Time.parse("15:44:33 2011-03-02")
 
-      @worker.extend(Module.new {
-        define_method(:puts) { |thing| last_puts = thing }
-      })
-
       @worker.very_verbose = true
       @worker.log("some log text")
 
-      assert_match /\[15:44:33 2011-03-02\] \d+: some log text/, last_puts
+      assert_match /\[15:44:33 2011-03-02\] \d+: some log text/, messages.string
     ensure
       Time.fake_time = nil
+      reset_logger
     end
   end
 
@@ -581,11 +641,30 @@ describe "Resque::Worker" do
         end
       end
 
-      @worker.very_verbose = true
-      stdout, stderr = capture_io { @worker.work(0) }
+      class DummyLogger
+        attr_reader :messages
 
-      assert_equal 3, stdout.scan(/retrying/).count
-      assert_equal 1, stdout.scan(/quitting/).count
+        def initialize
+          @messages = []
+        end
+
+        def info(message); @messages << message; end
+        alias_method :debug, :info
+        alias_method :warn,  :info
+        alias_method :error, :info
+        alias_method :fatal, :info
+      end
+
+      Resque.logger = DummyLogger.new
+      begin
+        @worker.work(0)
+        messages = Resque.logger.messages
+      ensure
+        reset_logger
+      end
+
+      assert_equal 3, messages.grep(/retrying/).count
+      assert_equal 1, messages.grep(/quitting/).count
     ensure
       class Redis::Client
         alias_method :reconnect, :original_reconnect
@@ -715,5 +794,33 @@ describe "Resque::Worker" do
         end
       end
     end
+  end
+
+  it "displays warning when using verbose" do
+    $warned_logger_severity_deprecation = false
+    stdout, stderr = capture_io { @worker.verbose }
+
+    assert stderr.match(/WARNING:/)
+  end
+
+  it "displays warning when using verbose=" do
+    $warned_logger_severity_deprecation = false
+    stdout, stderr = capture_io { @worker.verbose = true }
+
+    assert stderr.match(/WARNING:/)
+  end
+
+  it "displays warning when using very_verbose" do
+    $warned_logger_severity_deprecation = false
+    stdout, stderr = capture_io { @worker.very_verbose }
+
+    assert stderr.match(/WARNING:/)
+  end
+
+  it "displays warning when using very_verbose=" do
+    $warned_logger_severity_deprecation = false
+    stdout, stderr = capture_io { @worker.very_verbose = true }
+
+    assert stderr.match(/WARNING:/)
   end
 end